+++ packages/react-components/react-headless-components-preview/library/src/components/InteractionTag/InteractionTagSecondary/index.ts
@@ -1,0 +1,8 @@
+export { InteractionTagSecondary } from './InteractionTagSecondary';
+export { renderInteractionTagSecondary } from './renderInteractionTagSecondary';
+export { useInteractionTagSecondary } from './useInteractionTagSecondary';
+export type {
+  InteractionTagSecondarySlots,
+  InteractionTagSecondaryProps,
+  InteractionTagSecondaryState,
+} from './InteractionTagSecondary.types';