+++ packages/react-components/react-headless-components-preview/stories/src/Tags/TagGroup/index.stories.tsx
@@ -1,0 +1,19 @@
+import { TagGroup } from '@fluentui/react-headless-components-preview/tag-group';
+
+import descriptionMd from './TagGroupDescription.md';
+export { Default } from './TagGroupDefault.stories';
+export { Dismiss } from './TagGroupDismiss.stories';
+export { Disabled } from './TagGroupDisabled.stories';
+export { Select } from './TagGroupSelect.stories';
+
+export default {
+  title: 'Components/Tags/TagGroup',
+  component: TagGroup,
+  parameters: {
+    docs: {
+      description: {
+        component: descriptionMd,
+      },
+    },
+  },
+};