+++ packages/react-components/react-migration-v0-v9/library/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-migration-v0-v9
 
-This log was last generated on Tue, 26 May 2026 09:39:24 GMT and should not be manually modified.
+This log was last generated on Wed, 27 May 2026 11:18:57 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.6.30](https://github.com/microsoft/fluentui/tree/@fluentui/react-migration-v0-v9_v9.6.30)
+
+Wed, 27 May 2026 11:18:57 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-migration-v0-v9_v9.6.29..@fluentui/react-migration-v0-v9_v9.6.30)
+
+### Patches
+
+- Bump @fluentui/react-components to v9.74.1 ([commit](https://github.com/microsoft/fluentui/commit/5e0f816076baa63d07331287402ed4bd6d8e3c60) by beachball)
 
 ## [9.6.29](https://github.com/microsoft/fluentui/tree/@fluentui/react-migration-v0-v9_v9.6.29)